Refactor trends analysis to dedicated subpackage
Moves code specific to trends analysis to dedicated `trends`
submodule (maintains access via legacy endpoints).
