+++ tests/test_analysis.py
@@ -19,7 +19,8 @@ from pandas import Timestamp as T
 from pandas.testing import assert_series_equal, assert_frame_equal, assert_index_equal
 import pytest
 
-from market_analy import analysis, guis, trends, charts
+from market_analy import analysis, guis, charts
+from market_analy.trends import analy as analy_trends
 from market_analy.utils import UTC
 from market_analy.utils import bq_utils as bqu
 
@@ -1104,7 +1105,7 @@ class TestAnalysis:
         f = analy.trends_chart
         interval = "1D"
         kwargs = trend_period_kwargs
-        verify_app(f, trends.TrendsGui, interval, trend_kwargs, **kwargs)
+        verify_app(f, analy_trends.TrendsGui, interval, trend_kwargs, **kwargs)
 
         gui = f(interval, trend_kwargs, **kwargs, display=False)
         movements = analy.movements("1D", trend_kwargs, **kwargs)
@@ -1187,7 +1188,7 @@ class TestAnalysis:
             assert not controls.is_dark_single_case
             assert not controls.but_show_all.is_light
 
-        def assert_trend_reflects_move(move: trends.Movement):
+        def assert_trend_reflects_move(move: analy_trends.Movement):
             # verify gui as requried
             verify_controls_reflect_single_trend()
             assert (
@@ -1297,7 +1298,7 @@ class TestAnalysis:
         assert len(gui._rulers) == 2
 
         # verify break_rule
-        def verify_break_rule(rule: bqu.TrendRule, move: trends.Movement):
+        def verify_break_rule(rule: bqu.TrendRule, move: analy_trends.Movement):
             for c in rule.components:
                 assert c.visible
                 assert c in gui.chart.figure.marks
@@ -1322,7 +1323,7 @@ class TestAnalysis:
         verify_break_rule(break_rule, move)
 
         # verify limit_rule
-        def verify_limit_rule(rule: bqu.TrendRule, move: trends.Movement):
+        def verify_limit_rule(rule: bqu.TrendRule, move: analy_trends.Movement):
             for c in rule.components:
                 assert c.visible
                 assert c in gui.chart.figure.marks
